requirements/requirements.txt
Django==3.2.9
psycopg2-binary==2.9.2
django-braces==1.15.0
python-magic=0.4.24
uWSGI==2.0.20
